backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
firebase-admin==6.5.0
pytesseract==0.3.10
Pillow==10.3.0
PyPDF2==3.0.1
pdf2image==1.17.0
python-docx==1.1.2
lxml>=4.9.0,<5.3.0
spacy==3.7.4
pydantic==2.7.1
cryptography==42.0.7
httpx==0.27.0
pymupdf==1.24.2
